+++ collection_manager/collection_manager/entities/Collection.py
@@ -21,8 +21,7 @@ from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
 from fnmatch import fnmatch
-from glob import glob
-from typing import List, Optional
+from typing import Optional
 from urllib.parse import urlparse
 
 from collection_manager.entities.exceptions import MissingValueCollectionError
@@ -47,6 +46,7 @@ class Collection:
     forward_processing_priority: Optional[int] = None
     date_from: Optional[datetime] = None
     date_to: Optional[datetime] = None
+    preprocess: str = None
 
     @staticmethod
     def __decode_dimension_names(dimension_names_dict):
@@ -79,6 +79,8 @@ class Collection:
             date_to = datetime.fromisoformat(properties['to']) if 'to' in properties else None
             date_from = datetime.fromisoformat(properties['from']) if 'from' in properties else None
 
+            preprocess = json.dumps(properties['preprocess']) if 'preprocess' in properties else None
+
             collection = Collection(dataset_id=properties['id'],
                                     projection=properties['projection'],
                                     dimension_names=frozenset(Collection.__decode_dimension_names(properties['dimensionNames'])),
@@ -87,7 +89,8 @@ class Collection:
                                     historical_priority=properties['priority'],
                                     forward_processing_priority=properties.get('forward-processing-priority', None),
                                     date_to=date_to,
-                                    date_from=date_from)
+                                    date_from=date_from,
+                                    preprocess=preprocess)
             return collection
         except KeyError as e:
             raise MissingValueCollectionError(missing_value=e.args[0])