+++ collection_manager/collection_manager/services/CollectionProcessor.py
@@ -13,11 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 import logging
 import os.path
-from glob import glob
 from typing import Dict
-from datetime import datetime
 
 import yaml
 from collection_manager.entities import Collection
@@ -122,6 +121,10 @@ class CollectionProcessor:
             },
             'processors': CollectionProcessor._get_default_processors(collection)
         }
+
+        if collection.preprocess is not None:
+            config_dict['preprocess'] = json.loads(collection.preprocess)
+
         config_str = yaml.dump(config_dict)
         logger.debug(f"Templated dataset config:\n{config_str}")
         return config_str